Nest SQLite snapshots under backup
