+++ docs/cli/backup.md
@@ -1,7 +1,8 @@
 ---
-summary: "CLI reference for `openclaw backup` (create local backup archives)"
+summary: "CLI reference for `openclaw backup` (archives and SQLite-safe artifacts)"
 read_when:
   - You want a first-class backup archive for local OpenClaw state
+  - You need a SQLite-safe artifact for one OpenClaw database
   - You want to preview which paths would be included before reset or uninstall
 title: "Backup"
 ---
@@ -18,6 +19,11 @@ openclaw backup create --verify
 openclaw backup create --no-include-workspace
 openclaw backup create --only-config
 openclaw backup verify ./2026-03-09T08-00-00.000+08-00-openclaw-backup.tar.gz
+openclaw backup sqlite snapshot create --target global --repository ./snapshots
+openclaw backup sqlite snapshot create --agent main --repository ./snapshots
+openclaw backup sqlite snapshot list --repository ./snapshots
+openclaw backup sqlite snapshot verify ./snapshots/<snapshot-id>
+openclaw backup sqlite snapshot restore ./snapshots/<snapshot-id> --target ./restore/openclaw.sqlite
 ```
 
 ## Notes
@@ -27,7 +33,89 @@ openclaw backup verify ./2026-03-09T08-00-00.000+08-00-openclaw-backup.tar.gz
 - Existing archive files are never overwritten. Output paths inside the source state/workspace trees are rejected to avoid self-inclusion.
 - `openclaw backup verify <archive>` checks that the archive contains exactly one root manifest, rejects traversal-style archive paths, and confirms every manifest-declared payload exists in the tarball. `openclaw backup create --verify` runs that validation immediately after writing the archive.
 - `openclaw backup create --only-config` backs up just the active JSON config file.
-- If you only need a SQLite-safe sync artifact for one OpenClaw database, use [`openclaw snapshot`](/cli/snapshot). Backup is for broad recovery archives.
+- If you only need a SQLite-safe sync artifact for one OpenClaw database, use `openclaw backup sqlite snapshot`. `backup create` is for broad recovery archives.
+
+## SQLite snapshot artifacts
+
+Use `openclaw backup sqlite snapshot` when you need a syncable artifact for one
+SQLite database. A snapshot repository stores verified snapshot directories
+containing `manifest.json` and `database.sqlite`, so a host, container, object
+storage sync, or backup system can copy those files instead of copying a hot
+SQLite database.
+
+Use `openclaw backup create` when you need a broader local recovery archive for
+OpenClaw state, config, auth profiles, credentials, sessions, and optional
+workspaces. Backup archives may contain SQLite-safe database copies, but their
+output and restore model are archive-level, not a per-database snapshot
+repository.
+
+### What to sync
+
+Sync the snapshot directory created under the repository. A snapshot directory
+contains:
+
+- `manifest.json`
+- `database.sqlite`
+
+Do not sync live SQLite runtime files as the portability artifact:
+
+- `openclaw.sqlite`
+- `openclaw.sqlite-wal`
+- `openclaw.sqlite-shm`
+- `openclaw-agent.sqlite`
+- `openclaw-agent.sqlite-wal`
+- `openclaw-agent.sqlite-shm`
+
+Those files are hot runtime state. `openclaw backup sqlite snapshot create`
+reads the live database and writes a compact, verified SQLite artifact that can
+be copied by a host, container, object storage sync, or backup system.
+
+SQLite snapshot artifacts are still sensitive state. Global and per-agent
+artifacts can include auth profile records, session state, plugin state, and
+other credentials-adjacent data from the source SQLite database. Protect
+snapshot repositories with the same access controls, encryption, retention
+policy, and upload destination restrictions you use for OpenClaw backups and
+live state.
+
+### Named SQLite targets
+
+Use named targets when snapshotting OpenClaw-owned SQLite state:
+
+| Command                                                                     | Source                                         |
+| --------------------------------------------------------------------------- | ---------------------------------------------- |
+| `openclaw backup sqlite snapshot create --target global --repository <dir>` | Shared control-plane state database            |
+| `openclaw backup sqlite snapshot create --agent <id> --repository <dir>`    | Per-agent database for the normalized agent id |
+
+`--db <path>` remains available for explicit SQLite files and advanced scripts.
+Choose only one source selector: `--db`, `--target`, or `--agent`.
+
+Hosted runtimes should ask OpenClaw to materialize the named target instead of
+copying private SQLite paths. Hosts can sync only completed snapshot artifacts
+from the repository.
+
+### Restore a SQLite artifact
+
+Restore from the copied snapshot directory, not from the live source database
+files:
+
+```bash
+openclaw backup sqlite snapshot verify ./synced/snapshot
+openclaw backup sqlite snapshot restore ./synced/snapshot --target ./hydrated/openclaw.sqlite
+```
+
+Restore verifies the manifest, artifact hash, and SQLite integrity before
+copying the artifact to the target path. The target SQLite file must not already
+exist; stale `-wal`, `-shm`, and `-journal` sidecars at the target path are
+removed after the restore copy.
+
+### SQLite snapshot notes
+
+- Snapshot creation uses SQLite `VACUUM INTO`, so deleted-page remnants are not
+  carried into the artifact.
+- Snapshot repositories are local directories. Uploading or scheduling them is
+  intentionally left to the operator or a future integration.
+- This command does not add WAL bundle deltas, leases, failover automation, or
+  restore-on-boot behavior.
 
 ## What gets backed up
