+++ src/cli/program/command-registry-core.ts
@@ -83,11 +83,6 @@ const coreEntrySpecs: readonly CommandGroupDescriptorSpec<
         exportName: "registerBackupCommand",
       },
       {
-        commandNames: ["snapshot"],
-        loadModule: () => import("./register.snapshot.js"),
-        exportName: "registerSnapshotCommand",
-      },
-      {
         commandNames: ["migrate"],
         loadModule: () => import("./register.migrate.js"),
         exportName: "registerMigrateCommand",
